Separated movement into movement, whipcontrols, and torch controls
