+++ Assets/Scripts/Player Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90c542a48d8f7644787c5d1690d7148a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
