+++ Assets/Scripts/Player Scripts/TorchControls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9efb4d2de50e884c937f54c5fc06628
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
