+++ Assets/Scripts/Player Scripts/WhipControls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a9dfd8673b52b9418c3c80eda67c7a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
